Missing conversion

+++ android/src/main/java/com/swmansion/reanimated/layoutReanimation/SharedTransitionManager.java
@@ -444,7 +444,7 @@ public class SharedTransitionManager {
             preparedValues.put(key, value);
           } else {
             float pixelsValue = Utils.convertToFloat(value);
-            float dipValue = PixelUtil.toDIPFromPixel(pixelsValue);
+            double dipValue = PixelUtil.toDIPFromPixel(pixelsValue);
             preparedValues.put(key, dipValue);
           }
         }
